+++ Hollow/Hollow/Layers/Layer.h
@@ -3,7 +3,7 @@
 class Layer
 {
 public:
-	virtual void Update(float dt) = 0;
-	virtual void PreUpdate(float dt) = 0;
-	virtual void PostUpdate(float dt) = 0;
+	virtual void Update(double dt) = 0;
+	virtual void PreUpdate(double dt) = 0;
+	virtual void PostUpdate(double dt) = 0;
 };
